move ticket deletion from repository to job
